+++ extension/lib/shopify.api.class.js
@@ -35,21 +35,21 @@ class SGShopifyApi {
    */
   async addAddress (customerId, address) {
     return new Promise((resolve, reject) => {
-      this.postRequest(`/admin/customers/${customerId}/addresses.json`, {address}, (err, response) => {
+      this.postRequest(`/admin/customers/${customerId}/addresses.json`, { address }, (err, response) => {
         if (err) {
           // Some Shopify address validation error occurred
-          if (!Tools.isEmpty(response.errors)) {
+          if (err.code === 422) {
             const validationError = new FieldValidationError()
-            for (let path in response.errors) {
-              response.errors[path].forEach(message => {
-                validationError.addValidationMessage(path, message, address[path])
+            for (let fieldName in err.error) {
+              err.error[fieldName].forEach(message => {
+                validationError.addValidationMessage(fieldName, message, address[fieldName])
               })
             }
             return reject(validationError)
           }
           return reject(new UnknownError())
         }
-        return resolve({success: true})
+        return resolve({ success: true })
       })
     })
   }
@@ -85,24 +85,24 @@ class SGShopifyApi {
    */
   async updateAddress (customerId, address) {
     return new Promise((resolve, reject) => {
-      this.putRequest(`/admin/customers/${customerId}/addresses/${address.id}.json`, {address}, (err, response) => {
+      this.putRequest(`/admin/customers/${customerId}/addresses/${address.id}.json`, { address }, (err, response) => {
         if (err) {
           if (err.code === 404) {
             return reject(new InvalidCallError('Address not found'))
           }
           // Some Shopify address validation error occurred
-          if (!Tools.isEmpty(response.errors)) {
+          if (err.code === 422) {
             const validationError = new FieldValidationError()
-            for (let path in response.errors) {
-              response.errors[path].forEach(message => {
-                validationError.addValidationMessage(path, message, address[path])
+            for (let fieldName in err.error) {
+              err.error[fieldName].forEach(message => {
+                validationError.addValidationMessage(fieldName, message, address[fieldName])
               })
             }
             return reject(validationError)
           }
           return reject(new UnknownError())
         }
-        return resolve({success: true})
+        return resolve({ success: true })
       })
     })
   }
@@ -116,15 +116,17 @@ class SGShopifyApi {
     return new Promise((resolve, reject) => {
       this.putRequest(`/admin/customers/${customerId}/addresses/set.json?address_ids[]=${addressIds.join('&address_ids[]=')}&operation=destroy`, {}, (err, response) => {
         if (err) {
-          // Some Shopify address validation error occurred.
+          // Some Shopify address validation error occurred
+          if (err.code === 422) {
+            const validationError = new FieldValidationError()
+
+            if (err.error.match(/Cannot remove address ids because the default address id \(.*\) was included/)) {
+              return reject(new InvalidCallError('Cannot remove default address.'))
+            }
+          }
           return reject(new UnknownError())
         }
-
-        if (response.errors && response.errors.match(/Cannot remove address ids because the default address id \(.*\) was included/)) {
-          return reject(new InvalidCallError('Cannot remove default address.'))
-        }
-
-        return resolve({success: true})
+        return resolve({ success: true })
       })
     })
   }